Int til LONG

+++ Car Rental Service/CarApp+Gateway/Shared/Dto/Payment/PaymentDto.cs
@@ -5,6 +5,6 @@ public class PaymentDto
     public int BookingId { get; set; }
     public int CustomerId { get; set; }
     public string? ExpirationDate { get; set; } 
-    public int CreditCardRef { get; set; }
+    public long CreditCardRef { get; set; }
     public string? Status { get; set; }
 }
